+++ M.A.G.U.S/Bestiary/Undead/Vampire.cs
@@ -11,6 +11,7 @@ public sealed class Vampire : LivingDead
 {
     public Vampire()
     {
+        //Strength = 20;
         Occurrence = Occurrence.Rare;
         Size = Size.Human;
         AttackValue = 80;
@@ -18,15 +19,13 @@ public sealed class Vampire : LivingDead
         InitiateValue = 32;
         AttackModes =
         [
-            //new MeleeAttack(new BodyPart("Tooth1", GetDamage), AttackValue),
             new MeleeAttack(new BodyPart("Tooth", ThrowType._1D6), AttackValue),
-            //new MeleeAttack("Tooth3", AttackValue, () => GetDamage()),
             new MeleeAttack(new Longsword(), AttackValue)
         ];
-        HealthPoints = 40;
-        AstralMagicResistance = Byte.MaxValue;
-        MentalMagicResistance = Byte.MaxValue;
-        PoisonResistance = Byte.MaxValue;
+        HealthPoints = 40; // Regenerate 1 HP per turn
+        AstralMagicResistance = Int32.MaxValue;
+        MentalMagicResistance = Int32.MaxValue;
+        PoisonResistance = Int32.MaxValue;
         Intelligence = Enums.Intelligence.High;
         Alignment = Enums.Alignment.ChaosDeath;
         ExperiencePoints = 4000;
